requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-SocketIO>=5.3,<6.0
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
simple-websocket>=1.1,<2.0
pytest>=8.0,<9.0
